feat(charts): export ChartContainer + add custom DEFAULT_CHARTS_COLOR_SCHEME
